Services and repositories

+++ backend/src/main/java/com/example/backend/dto/response/cart/CartItemResponse.java
@@ -1,5 +1,6 @@
 package com.example.backend.dto.response.cart;
 
+import com.example.backend.dto.response.catalog.ProductResponse;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -15,5 +16,5 @@ public class CartItemResponse {
     private Long cartItemId;
     private ProductResponse product; // nested
     private int quantity;
-    private BigDecimal subtotal;     // price × qty
+    private Double subtotal;     // price × qty
 }
